Move Sample and Upscaling to their own files
