+++ profiling/src/profile/internal/mod.rs
@@ -5,16 +5,20 @@ mod function;
 mod label;
 mod location;
 mod mapping;
+mod sample;
 mod stack_trace;
 mod string;
+mod upscaling;
 mod value_type;
 
 pub use function::*;
 pub use label::*;
 pub use location::*;
 pub use mapping::*;
+pub use sample::*;
 pub use stack_trace::*;
 pub use string::*;
+pub use upscaling::*;
 pub use value_type::*;
 
 use super::pprof;
